Split tensor_generator from iter_batches into separate modules
