+++ tiledb/ml/readers/tensorflow.py
@@ -8,8 +8,8 @@ import tensorflow as tf
 
 import tiledb
 
-from ._batch_utils import SparseTileDBTensorGenerator, tensor_generator
 from ._buffer_utils import get_attr_names, get_buffer_size
+from ._tensor_gen import SparseTileDBTensorGenerator, tensor_generator
 
 # TODO: We have to track the following issues:
 # - https://github.com/tensorflow/tensorflow/issues/47532
